+++ realm/realm-library/src/androidTest/java/io/realm/entities/IOSChild.java
@@ -20,7 +20,7 @@ import io.realm.RealmObject;
 import io.realm.annotations.Required;
 
 public class IOSChild extends RealmObject {
-    @Required
+
     private String name;
 
     public String getName() {